requirements.txt
absl-py==0.8.1
blis==0.7.4
catalogue==2.0.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
cymem==2.0.5
Cython==0.29.23
cytoolz==0.9.0.1
decorator==4.4.2
dill==0.3.4
easydict==1.9
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
-e git+https://github.com/UchihaLake/python2.git@f87273611c15f9fb29b30cca8a60e68a12d67891#egg=faster_rcnn&subdirectory=mlprac\cv\m2trans\fasterrcnn_vg\lib
future==0.17.1
grpcio==1.25.0
h5py==2.8.0
idna==2.10
imageio==2.9.0
Jinja2==3.0.1
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==2.0.1
matplotlib==2.2.3
msgpack==0.6.2
multiprocess==0.70.9
murmurhash==1.0.5
networkx==2.5.1
numpy==1.16.4
opencv-python==4.5.2.54
packaging==20.9
pathlib==1.0.1
pathos==0.2.3
pathy==0.5.2
Pillow==8.2.0
plac==0.9.6
pox==0.2.7
ppft==1.6.6.1
preshed==3.0.5
protobuf==3.10.0
pycocotools==2.0.2
pycparser==2.20
pydantic==1.7.4
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.4.1
regex==2017.4.5
requests==2.25.1
scikit-image==0.18.1
scipy==1.6.3
six==1.16.0
smart-open==3.0.0
spacy==3.0.6
spacy-legacy==3.0.6
srsly==2.4.1
tensorboard==1.14.0
tensorboardX==2.2
thinc==8.0.5
tifffile==2021.6.14
toolz==0.11.1
torch==1.9.0
torchvision==0.10.0
tqdm==4.32.2
typer==0.3.2
typing-extensions==3.10.0.0
ujson==1.35
urllib3==1.26.5
wasabi==0.8.2
Werkzeug==0.16.0
wrapt==1.10.11
zipp==3.4.1
